+++ src/grafx/processors/core/fft_filterbank.py
@@ -1,7 +1,8 @@
-import math
 import warnings
 
 import torch
+
+from grafx.processors.core.scale import bark_to_hz
 
 
 def _create_triangular_filterbank(
@@ -28,146 +29,6 @@ def _create_triangular_filterbank(
     fb = torch.max(zero, torch.min(down_slopes, up_slopes))
 
     return fb
-
-
-def _hz_to_bark(freqs: float, bark_scale: str = "traunmuller") -> float:
-    r"""Convert Hz to Barks.
-
-    Args:
-        freqs (float): Frequencies in Hz
-        bark_scale (str, optional): Scale to use: ``traunmuller``, ``schroeder`` or ``wang``. (Default: ``traunmuller``)
-
-    Returns:
-        barks (float): Frequency in Barks
-    """
-
-    if bark_scale not in ["schroeder", "traunmuller", "wang"]:
-        raise ValueError(
-            'bark_scale should be one of "schroeder", "traunmuller" or "wang".'
-        )
-
-    if bark_scale == "wang":
-        return 6.0 * math.asinh(freqs / 600.0)
-    elif bark_scale == "schroeder":
-        return 7.0 * math.asinh(freqs / 650.0)
-    # Traunmuller Bark scale
-    barks = ((26.81 * freqs) / (1960.0 + freqs)) - 0.53
-    # Bark value correction
-    if barks < 2:
-        barks += 0.15 * (2 - barks)
-    elif barks > 20.1:
-        barks += 0.22 * (barks - 20.1)
-
-    return barks
-
-
-def _bark_to_hz(barks: torch.Tensor, bark_scale: str = "traunmuller") -> torch.Tensor:
-    """Convert bark bin numbers to frequencies.
-
-    Args:
-        barks (torch.Tensor): Bark frequencies
-        bark_scale (str, optional): Scale to use: ``traunmuller``,``schroeder`` or ``wang``. (Default: ``traunmuller``)
-
-    Returns:
-        freqs (torch.Tensor): Barks converted in Hz
-    """
-
-    if bark_scale not in ["schroeder", "traunmuller", "wang"]:
-        raise ValueError(
-            'bark_scale should be one of "traunmuller", "schroeder" or "wang".'
-        )
-
-    if bark_scale == "wang":
-        return 600.0 * torch.sinh(barks / 6.0)
-    elif bark_scale == "schroeder":
-        return 650.0 * torch.sinh(barks / 7.0)
-    # Bark value correction
-    if any(barks < 2):
-        idx = barks < 2
-        barks[idx] = (barks[idx] - 0.3) / 0.85
-    elif any(barks > 20.1):
-        idx = barks > 20.1
-        barks[idx] = (barks[idx] + 4.422) / 1.22
-
-    # Traunmuller Bark scale
-    freqs = 1960 * ((barks + 0.53) / (26.28 - barks))
-
-    return freqs
-
-
-def _hz_to_mel(freq: float, mel_scale: str = "htk") -> float:
-    r"""Convert Hz to Mels.
-
-    Args:
-        freqs (float): Frequencies in Hz
-        mel_scale (str, optional): Scale to use: ``htk`` or ``slaney``. (Default: ``htk``)
-
-    Returns:
-        mels (float): Frequency in Mels
-    """
-
-    if mel_scale not in ["slaney", "htk"]:
-        raise ValueError('mel_scale should be one of "htk" or "slaney".')
-
-    if mel_scale == "htk":
-        return 2595.0 * math.log10(1.0 + (freq / 700.0))
-
-    # Fill in the linear part
-    f_min = 0.0
-    f_sp = 200.0 / 3
-
-    mels = (freq - f_min) / f_sp
-
-    # Fill in the log-scale part
-    min_log_hz = 1000.0
-    min_log_mel = (min_log_hz - f_min) / f_sp
-    logstep = math.log(6.4) / 27.0
-
-    if freq >= min_log_hz:
-        mels = min_log_mel + math.log(freq / min_log_hz) / logstep
-
-    return mels
-
-
-def _mel_to_hz(mels: torch.Tensor, mel_scale: str = "htk") -> torch.Tensor:
-    """Convert mel bin numbers to frequencies.
-
-    Args:
-        mels (Tensor): Mel frequencies
-        mel_scale (str, optional): Scale to use: ``htk`` or ``slaney``. (Default: ``htk``)
-
-    Returns:
-        freqs (Tensor): Mels converted in Hz
-    """
-
-    if mel_scale not in ["slaney", "htk"]:
-        raise ValueError('mel_scale should be one of "htk" or "slaney".')
-
-    if mel_scale == "htk":
-        return 700.0 * (10.0 ** (mels / 2595.0) - 1.0)
-
-    # Fill in the linear scale
-    f_min = 0.0
-    f_sp = 200.0 / 3
-    freqs = f_min + f_sp * mels
-
-    # And now the nonlinear scale
-    min_log_hz = 1000.0
-    min_log_mel = (min_log_hz - f_min) / f_sp
-    logstep = math.log(6.4) / 27.0
-
-    log_t = mels >= min_log_mel
-    freqs[log_t] = min_log_hz * torch.exp(logstep * (mels[log_t] - min_log_mel))
-
-    return freqs
-
-
-def _hz_to_log(freqs):
-    return math.log(freqs)
-
-
-def _log_to_hz(logs):
-    return torch.exp(logs)
 
 
 def fft_triangular_filterbank(